Upload of Loan Visualization
The objective of this project is to provide valuable guidance for stakeholders in optimizing lending strategies, risk assessment, and customer engagement. Analysis of Prosper Loan lending services using pandas, NumPy, and data visualization tools, Seaborn and Matplotlib, was performed to gain a comprehensive understanding of Prosper Loan lending services' performance and customer trends.